refactor(controllers): extract HelpSessionProvisioner from HelpSessionController
- Move provisionHelpSession, provisionFailureKind, buildHelpEnv,
  revokeHelpSession, asStringRecord, and HelpSessionRef into a sibling
  module — pure functions with no controller state, no behavior change
- Controller imports these by relative path (no barrel file)
