+++ src/controllers/HelpSessionController.ts
@@ -23,6 +23,14 @@ import { buildResumeCommand, buildResumeLatestCommand } from "@shared/types/agen
 import { resolveDaintreeMcpTier } from "@shared/types/project";
 import type { SnapshotInfo } from "@shared/types/ipc/git";
 import { isAssistantOnlyAgentId } from "@shared/config/agentIds";
+import {
+  type HelpSessionRef,
+  provisionHelpSession,
+  provisionFailureKind,
+  buildHelpEnv,
+  revokeHelpSession,
+  asStringRecord,
+} from "./HelpSessionProvisioner";
 
 const HIBERNATE_VALID_MINUTES: readonly number[] = [0, 5, 15, 30, 60, 120];
 const DEFAULT_HIBERNATE_MINUTES = 5;
@@ -281,73 +289,6 @@ export interface HelpLaunchOptions {
   resumeOnly?: boolean;
 }
 
-interface HelpSessionRef {
-  sessionId: string;
-  sessionPath: string;
-  token: string;
-  mcpUrl: string | null;
-  windowId: number;
-}
-
-/**
- * Per-agent env injection for help-session launches. Today this is a
- * placeholder shape — no agent currently requires renderer-side env beyond
- * the universal `DAINTREE_MCP_TOKEN` / `DAINTREE_WINDOW_ID` set in
- * `buildHelpEnv`. The hook stays so future per-agent env additions have a
- * single place to land.
- */
-function agentSpawnEnv(_agentId: string, _sessionPath: string): Record<string, string> {
-  return {};
-}
-
-type ProvisionFailureCode =
-  | "MCP_NOT_READY"
-  | "MCP_SERVER_NOT_STARTED"
-  | "MCP_PROBE_FAILED"
-  | "UNKNOWN";
-
-type ProvisionOutcome =
-  | { ok: true; session: HelpSessionRef }
-  | { ok: false; code: ProvisionFailureCode; message: string };
-
-async function provisionHelpSession(
-  project: HelpProjectRef,
-  agentId: string,
-  context?: ActionContext
-): Promise<ProvisionOutcome> {
-  try {
-    const result = await window.electron.help.provisionSession({
-      projectId: project.id,
-      projectPath: project.path,
-      agentId,
-      ...(context && { context }),
-    });
-    if (!result) {
-      return {
-        ok: false,
-        code: "UNKNOWN",
-        message: "Couldn't provision help session.",
-      };
-    }
-    return { ok: true, session: result };
-  } catch (err) {
-    logError("Failed to provision help session", err);
-    const code =
-      err && typeof err === "object" && "code" in err
-        ? (err as Record<string, unknown>).code
-        : undefined;
-    const message = formatErrorMessage(err, "Couldn't provision help session");
-    if (
-      code === "MCP_SERVER_NOT_STARTED" ||
-      code === "MCP_PROBE_FAILED" ||
-      code === "MCP_NOT_READY"
-    ) {
-      return { ok: false, code, message };
-    }
-    return { ok: false, code: "UNKNOWN", message };
-  }
-}
-
 /**
  * Three-state version probe.
  * - `ok`: version meets the minimum (or no minimum is configured).
@@ -364,20 +305,6 @@ type VersionProbeResult =
   | { status: "ok" }
   | { status: "indeterminate" }
   | { status: "too-old"; block: VersionTooOld };
-
-function provisionFailureKind(code: ProvisionFailureCode): LaunchErrorKind {
-  switch (code) {
-    case "MCP_SERVER_NOT_STARTED":
-      return "mcp-server-not-started";
-    // Legacy `MCP_NOT_READY` errors fall through to the probe-failed shape —
-    // the "server responded badly" copy is the closer fit.
-    case "MCP_PROBE_FAILED":
-    case "MCP_NOT_READY":
-      return "mcp-probe-failed";
-    default:
-      return "spawn-failed";
-  }
-}
 
 // `refresh=true` bypasses the 12h AgentVersionService cache — pass on retry
 // so a user who manually updates the CLI outside Daintree's update flow can
@@ -445,39 +372,6 @@ export async function loadCustomLaunchFlags(): Promise<string[]> {
     logError("Failed to load helpAssistant launch flags", err);
     return [];
   }
-}
-
-function buildHelpEnv(
-  session: HelpSessionRef | null,
-  projectId: string | null,
-  agentId: string
-): Record<string, string> | undefined {
-  if (!session) return undefined;
-  const env: Record<string, string> = {
-    DAINTREE_MCP_TOKEN: session.token,
-    DAINTREE_WINDOW_ID: String(session.windowId),
-    ...agentSpawnEnv(agentId, session.sessionPath),
-  };
-  if (session.mcpUrl) env.DAINTREE_MCP_URL = session.mcpUrl;
-  if (projectId) env.DAINTREE_PROJECT_ID = projectId;
-  return env;
-}
-
-function revokeHelpSession(sessionId: string | null): void {
-  if (!sessionId) return;
-  window.electron.help.revokeSession(sessionId).catch((err) => {
-    logError("Failed to revoke help session", err);
-  });
-}
-
-function asStringRecord(value: unknown): Record<string, string> | undefined {
-  if (!value || typeof value !== "object") return undefined;
-  const out: Record<string, string> = {};
-  for (const [k, v] of Object.entries(value)) {
-    if (typeof v !== "string") return undefined;
-    out[k] = v;
-  }
-  return out;
 }
 
 function notifyLaunchFailed(agentId: string, reason: string): void {
